+++ packages/core/src/footer/styles/RightWrapper.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import { mq } from '@heathmont/moon-utils';
+
+const RightWrapper = styled.div(({ theme: { breakpoint } }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  [mq(breakpoint.medium)]: {
+    flexDirection: 'row-reverse',
+  },
+}));
+
+export default RightWrapper;