+++ packages/core/src/header/Header.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import styled from 'styled-components';
+import { rem } from '@heathmont/moon-utils';
+
+import Container from './styles/Container';
+import RightWrapper from './styles/RightWrapper';
+
+export interface HeaderProps {
+  backButton?: JSX.Element;
+  closeButton?: JSX.Element;
+  icons?: JSX.Element[];
+  isDivider?: boolean;
+}
+
+const Label = styled.div({
+  paddingTop: rem(4),
+});
+
+const IconWrapper = styled.span({
+  marginLeft: rem(8),
+});
+
+const BackButtonIcon = styled.div({
+  cursor: 'pointer',
+  marginRight: rem(8),
+  minWidth: rem(32),
+});
+
+const ControlsCloseIcon = styled.div({
+  cursor: 'pointer',
+  marginLeft: rem(8),
+});
+
+const Header: React.FC<HeaderProps> = ({
+  children,
+  backButton,
+  closeButton,
+  icons,
+  isDivider,
+}) => (
+  <Container isDivider={isDivider}>
+    {backButton && <BackButtonIcon>{backButton}</BackButtonIcon>}
+    <Label>{children}</Label>
+    <RightWrapper>
+      {icons &&
+        icons.map((item, index) => (
+          <IconWrapper key={`hicon-${index}`}>{item}</IconWrapper>
+        ))}
+      {closeButton && <ControlsCloseIcon>{closeButton}</ControlsCloseIcon>}
+    </RightWrapper>
+  </Container>
+);
+
+export default Header;