+++ packages/core/src/header/styles/RightWrapper.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { rem } from '@heathmont/moon-utils';
+
+const RightWrapper = styled.div({
+  display: 'flex',
+  alignItems: 'center',
+  fontSize: rem(32),
+});
+
+export default RightWrapper;