+++ packages/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 7.1.0
+
+### Minor Changes
+
+- Introduce Header & Footer
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-icons@7.1.0
+  - @heathmont/moon-themes@7.1.0
+  - @heathmont/moon-utils@7.1.0
+
 ## 7.0.0
 
 ### Major Changes